+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Don't strip query parameters when converting key ID to actor ID.
 - Improved error handling during verification of GET requests to portable inboxes and outboxes.
 - Log error details when verifying activity fetch requests.
+- Return empty collection if request to portable outbox is not signed.
 
 ### Fixed